requirements.txt
aniso8601==9.0.1
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
idna==3.3
ipfs-api==0.2.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
protobuf==3.19.0
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
